safari-534.53-branch/LayoutTests/svg/dynamic-updates/SVGFETurbulenceElement-svgdom-baseFrequency-prop-expected.txt

CONSOLE MESSAGE: line 39: TypeError: 'undefined' is not an object (evaluating 'turbulence.baseFrequencyX.baseVal = 0.05')
FAIL: Timed out waiting for notifyDone to be called
SVG 1.1 dynamic update tests

Tests dynamic updates of the 'baseFrequency' property of the SVGFETurbulenceElment object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL Math.round(turbulence.baseFrequencyX.baseVal * 1000) should be 70. Threw exception TypeError: 'undefined' is not an object (evaluating 'turbulence.baseFrequencyX.baseVal')
FAIL Math.round(turbulence.baseFrequencyY.baseVal * 1000) should be 70. Threw exception TypeError: 'undefined' is not an object (evaluating 'turbulence.baseFrequencyY.baseVal')
